Remove obsolete identifiers in CMS data.
