+++ src/components/Head/index.tsx
@@ -1,29 +1,9 @@
 import React, { FunctionComponent } from 'react'
 import Helmet from 'react-helmet'
 import { useLocale } from '../../utils/hooks'
-import { TranslationCollection } from '../../cms/i18n-lib'
+import { Head as HeadProps } from '../../data/head/types'
 
-export interface Head {
-  title: string
-  meta: MetaContent
-}
-
-export interface MetaContent {
-  description: string
-  keywords: string
-}
-
-interface MetaContentLocalized {
-  description: TranslationCollection<string>
-  keywords: TranslationCollection<string>
-}
-
-export interface HeadLocalized {
-  title: TranslationCollection<string>
-  meta: MetaContentLocalized
-}
-
-export const Head: FunctionComponent<Head> = ({ title, meta }) => {
+export const Head: FunctionComponent<HeadProps> = ({ title, meta }) => {
   const locale = useLocale()
   return (
     <Helmet
