+++ src/components/layout.tsx
@@ -2,9 +2,10 @@ import React, { FunctionComponent } from 'react'
 import { Navbar } from './Navbar'
 import { Footer } from './Footer'
 import { Link } from 'gatsby'
+import { Head as HeadProps } from '../data/head/types'
 import { Head } from './Head'
 
-const Layout: FunctionComponent<Head> = ({ children, title, meta }) => (
+const Layout: FunctionComponent<HeadProps> = ({ children, title, meta }) => (
   <div>
     <Head title={title} meta={meta} />
     <header>
