+++ src/data/faqEntry/faq.json
@@ -1,5 +1,4 @@
 {
-  "id": "FAQ Page",
   "headline": [
     {
       "locale": "en",
@@ -61,7 +60,6 @@
     }
   ],
   "head": {
-    "id": "Head",
     "title": [
       {
         "locale": "en",
