+++ src/data/faqEntry/types.ts
@@ -1,4 +1,4 @@
-import { Head, HeadLocalized } from '../../components/Head'
+import { Head, HeadLocalized } from '../head/types'
 import { TranslationCollection as TranslationCollectionOld } from '../types'
 import { TranslationCollection } from '../../cms/i18n-lib'
 
@@ -10,7 +10,6 @@ export interface FAQPage {
 }
 
 export type FAQPageOnDisk = {
-  id: string
   head: HeadLocalized
   faqEntries: TranslationCollectionOld<FAQEntry>[]
 } & {
