+++ src/data/home/home.json
@@ -147,7 +147,6 @@
     }
   ],
   "head": {
-    "id": "Head",
     "title": [
       {
         "locale": "en",
