+++ src/data/imprint/imprint.json
@@ -19,7 +19,6 @@
     }
   ],
   "head": {
-    "id": "Imprint",
     "title": [
       {
         "locale": "en",
